mbus-serial: Kludge around i.MX28 AUART delays.
On the Freescale i.MX286 SoC, the mxs_auart driver suffers from
significant delays when reading at low baud rates.  The underlying cause
of this is still under investigation but in the meantime, in order to
get successful reads at the 1200 baud and above rates, it is necessary
to add a pause before each read.

This will need extension if sub-1200 baud rates are required.